backend/requirements.txt
fastapi
uvicorn
pandas
numpy
tensorflow
faiss-cpu
thefuzz[speedup]
requests
scikit-learn